Crimes reported in Abilene, Texas

Format

A data frame with 8 observations on the following 5 variables.

Crime

a factor with levels Aggravated assault Arson Burglary Forcible rape Larceny theft Murder Robbery Vehicle theft

X1992

a numeric vector

X92percent

a numeric vector

X1999

a numeric vector

X99percent

a numeric vector

Source

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury ======= A data frame with 16 observations on three variables.

crimetype

a factor with levels Aggravated assault, Arson, Burglary, Forcible rape, Larceny theft, Murder, Robbery, and Vehicle theft.

year

a factor with levels 1992 and 1999

number

number of reported crimes

Abilene >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Data used in Exercise 1.23 and 2.79

<<<<<<< HEAD data(Abilene) attach(Abilene) par(mfrow=c(2,1)) barplot(X1992,names.arg=c("Murder","Rape","Robbery","Assault","Burglary", "Larceny","V.Theft","Arson"),col="blue",main="Crime 1992") barplot(X1999,names.arg=c("Murder","Rape","Robbery","Assault","Burglary", "Larceny","V.Theft","Arson"),col="red",main="Crime 1999") par(mfrow=c(1,1)) detach(Abilene)

======= par(mfrow = c(2, 1)) barplot(Abilene$number[Abilene$year=="1992"], names.arg = Abilene$crimetype[Abilene$year == "1992"], main = "1992 Crime Stats", col = "red") barplot(Abilene$number[Abilene$year=="1999"], names.arg = Abilene$crimetype[Abilene$year == "1999"], main = "1999 Crime Stats", col = "blue") par(mfrow = c(1, 1))

## Not run: ------------------------------------ # library(ggplot2) # ggplot(data = Abilene, aes(x = crimetype, y = number, fill = year)) + # geom_bar(stat = "identity", position = "dodge") + # theme_bw() + # theme(axis.text.x = element_text(angle = 30, hjust = 1)) ## ---------------------------------------------

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets