Abortion rate by region of country

Format

A data frame with 51 observations on the following 9 variables. ======= A data frame with 51 observations on the following 10 variables. >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

state

a factor with levels alabama alaska arizona arkansas california colorado connecticut delaware dist of columbia florida georgia hawaii idaho illinois indiana iowa kansas kentucky louisiana maine maryland massachusetts michigan minnesota mississippi missouri montana nebraska nevada new hampshire new jersey new mexico new york north carolina north dakota ohio oklahoma oregon pennsylvania rhode island south carolina south dakota tennessee texas utah vermont virginia washington west virginia wisconsin wyoming

region

a factor with levels midwest northeast south west

regcode

a numeric vector
<<<<<<< HEAD
X88rate

a numeric vector

X92rate

a numeric vector

X96rate

a numeric vector

X88provid

a numeric vector

X92provid

a numeric vector

lowhigh

a numeric vector
=======
rate1988

a numeric vector

rate1992

a numeric vector

rate1996

a numeric vector

provide1988

a numeric vector

provide1992

a numeric vector

lowhigh

a numeric vector

rate

a factor with levels Low and High
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Abortion

Data used in Exercise 8.51

<<<<<<< HEAD attach(Abortion) AbortionRate <- cut(X96rate,breaks=c(0,20,10000) ) levels(AbortionRate) <- c("Low","High") table(region,AbortionRate) chisq.test(table(region,AbortionRate)) detach(Abortion)

======= T1 <- xtabs(~region + rate, data = Abortion) T1 chisq.test(T1)

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets