Math achievement test scores by gender for 25 high school students

Data used in Example 7.14 and Exercise 10.7

Achieve

Format

A data frame/tibble with 25 observations on two variables

score

mathematics achiement score

gender

a factor with 2 levels boys and girls

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.

Examples

anova(lm(score ~ gender, data = Achieve))

#> Analysis of Variance Table
#> 
#> Response: score
#>           Df  Sum Sq Mean Sq F value  Pr(>F)  
#> gender     1  403.85  403.85  4.4271 0.04651 *
#> Residuals 23 2098.15   91.22                  
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

t.test(score ~ gender, var.equal = TRUE, data = Achieve)

#> 
#> 	Two Sample t-test
#> 
#> data:  score by gender
#> t = -2.1041, df = 23, p-value = 0.04651
#> alternative hypothesis: true difference in means between group boys and group girls is not equal to 0
#> 95 percent confidence interval:
#>  -15.9543968  -0.1353468
#> sample estimates:
#>  mean in group boys mean in group girls 
#>            76.41667            84.46154 
#>