Median costs of appendectomies at three different types of North Carolina

Format

A data frame with 59 observations on the following 5 variables.

Cost

a numeric vector

Region

a numeric vector

Rural

a numeric vector

Regional

a numeric vector

Metropol

a numeric vector
======= A data frame with 59 observations on the following two variables.

cost

median costs of appendectomies at hospitals across the state of North Carolina in 1992

region

a numeric vector classifying each hospital (1 = rural), (2 = regional), (3 = metropolitan)
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Appendec

Data for Exercise 10.60

str(Appendec) <<<<<<< HEAD attach(Appendec) boxplot(Cost~Region) anova(lm(Cost~as.factor(Region))) detach(Appendec)

======= Appendec$regionc <- ifelse(Appendec$region == 1, "rural", ifelse(Appendec$region == 2, "regional", "metropolitan")) boxplot(cost ~ regionc, data = Appendec, col = c("red", "blue", "cyan")) anova(lm(cost ~ as.factor(regionc), data = Appendec))

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets