Number of defective auto gears produced by two manufacturers

Data for Exercise 7.46

Autogear

Format

A data frame/tibble with 20 observations on two variables

defectives

number of defective gears in the production of 100 gears per day

manufacturer

a factor with levels A and B

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.

Examples

t.test(defectives ~ manufacturer, data = Autogear)
#> 
#> 	Welch Two Sample t-test
#> 
#> data:  defectives by manufacturer
#> t = -2.2465, df = 37.587, p-value = 0.03063
#> alternative hypothesis: true difference in means between group A and group B is not equal to 0
#> 95 percent confidence interval:
#>  -9.4122347 -0.4877653
#> sample estimates:
#> mean in group A mean in group B 
#>           23.80           28.75 
#> 
wilcox.test(defectives ~ manufacturer, data = Autogear)
#> Warning: cannot compute exact p-value with ties
#> 
#> 	Wilcoxon rank sum test with continuity correction
#> 
#> data:  defectives by manufacturer
#> W = 114.5, p-value = 0.02115
#> alternative hypothesis: true location shift is not equal to 0
#> 
t.test(defectives ~ manufacturer, var.equal = TRUE, data = Autogear)
#> 
#> 	Two Sample t-test
#> 
#> data:  defectives by manufacturer
#> t = -2.2465, df = 38, p-value = 0.03056
#> alternative hypothesis: true difference in means between group A and group B is not equal to 0
#> 95 percent confidence interval:
#>  -9.4106236 -0.4893764
#> sample estimates:
#> mean in group A mean in group B 
#>           23.80           28.75 
#>