Basic Statistics and Data Analysis

Data and functions for the book Basic Statistics and Data Analysis

Details

Package:

PASWR

Type:

Package

Version:

1.02

Date:

2016-02-25

License:

GPL (>=2)

The package BSDA provides data and functions for the book Basic Statistics and Data Analysis

References

Kitchens, L. J (2003) Basic Statistics and Data Analysis. Brooks/Cole.

Author
Alan T. Arnholt Maintainer: <arnholtat@appstate.edu>