Graduation rates for student athletes and nonathletes in the Big Ten Conf.

Format

A data frame with 11 observations on the following 5 variables.

School

a factor with levels Illinois Indiana Iowa Michigan Michigan State Minnesota Northwestern Ohio State Penn State Purdue Wisconsin

X1984.85students

a numeric vector

X1984.85athletes

a numeric vector

X1993.94students

a numeric vector

X1993.94athletes

a numeric vector
======= A data frame with 44 observations on the following four variables.

school

a factor with levels Illinois Indiana Iowa Michigan Michigan State Minnesota Northwestern Ohio State Penn State Purdue Wisconsin

rate

graduation rate

year

factor with two levels 1984-1985 and 1993-1994

status

factor with two levels athlete and student
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Bigten

Data for Exercises 1.124 and 2.94

<<<<<<< HEAD str(Bigten) attach(Bigten) boxplot(X1993.94students,X1993.94athletes,names=c("Students","Athletes"), ylab="1993-1994 Graduation Rates") plot(X1993.94students,X1993.94athletes,xlab="1993-1994 students", ylab="1993-1994 athletes") detach(Bigten)

======= boxplot(rate ~ status, data = subset(Bigten, year = "1993-1994"), horizontal = TRUE, main = "Graduation Rates 1993-1994") with(data = Bigten, tapply(rate, list(year, status), mean) )

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets