Brain weight versus body weight of 28 animals

Format

A data frame with 28 observations on the following 5 variables.

Species

a factor with levels Africian elephant Asian Elephant Brachiosaurus Cat Chimpanzee Cow Diplodocus Donkey Giraffe Goat Gorilla Gray wolf Guinea Pig Hamster Horse Human Jaguar Kangaroo Mole Mouse Mt Beaver Pig Potar monkey Rabbit Rat Rhesus monkey Sheep Triceratops

body.wt

a numeric vector

brain.wt

a numeric vector

logbody

a numeric vector

logbrain

a numeric vector
======= A data frame with 28 observations on the following three variables.

species

a factor with levels African elephant, Asian Elephant, Brachiosaurus, Cat, Chimpanzee, Cow, Diplodocus, Donkey, Giraffe, Goat, Gorilla, Gray wolf, Guinea Pig, Hamster, Horse, Human, Jaguar, Kangaroo, Mole, Mouse, Mt Beaver, Pig, Potar monkey, Rabbit, Rat, Rhesus monkey, Sheep, and Triceratops

bodyweight

body weight (in kg)

brainweight

brain weight (in g)
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Brain

Data for Exercises 2.15, 2.44, 2.58 and Examples 2.3 and 2.20

<<<<<<< HEAD str(Brain) attach(Brain) plot(logbody,logbrain,pch=19,col="blue",main="Example 2.3") model <- lm(logbrain~logbody) abline(model) detach(Brain)

======= plot(log(brainweight) ~ log(bodyweight), data = Brain, pch = 19, col = "blue", main = "Example 2.3") mod <- lm(log(brainweight) ~ log(bodyweight), data = Brain) abline(mod, lty = "dashed", col = "blue")

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets