Repair costs of vehicles crashed into a barrier at 5 miles per hour

Format

A data frame with 23 observations on the following 2 variables.

Car

a factor with levels Buick Century Buick Skylark Chevrolet Cavalier Chevrolet Corsica Chevrolet Lumina Dodge Dynasty Dodge Monaco Ford Taurus Ford Tempo Honda Accord Hyundai Sonata Mazda 626 Mitsubishi Galant Nissan Stanza Oldsmobile Calais Oldsmobile Ciere Plymouth Acclaim Pontiac 6000 Pontiac Grand Am Pontiac Sunbird Saturn SL2 Subaru Legacy Toyota Camry

repair

a numeric vector
======= A data frame with 23 observations on the following two variables.

car

a factor with levels Buick Century, Buick Skylark, Chevrolet Cavalier, Chevrolet Corsica, Chevrolet Lumina, Dodge Dynasty, Dodge Monaco, Ford Taurus, Ford Tempo, Honda Accord, Hyundai Sonata, Mazda 626, Mitsubishi Galant, Nissan Stanza, Oldsmobile Calais, Oldsmobile Ciere, Plymouth Acclaim, Pontiac 6000, Pontiac Grand Am, Pontiac Sunbird, Saturn SL2, Subaru Legacy, and Toyota Camry

repair

total repair cost (in dollars) after crashing a car into a barrier four times while the car was traveling at 5 miles per hour
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Bumpers

Data for Exercise 1.73

<<<<<<< HEAD str(Bumpers) attach(Bumpers) EDA(repair) sum(repair>(mean(repair)-sd(repair)) & repair < (mean(repair)+sd(repair)))/length(repair) stripchart(repair,method="stack") library(lattice) dotplot(Car~repair) detach(Bumpers)

======= EDA(Bumpers$repair) stripchart(Bumpers$repair, method = "stack", pch = 19, col = "blue") library(lattice) dotplot(car ~ repair, data = Bumpers)

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets