Median charges for coronary bypass at 17 hospitals in North Carolina

Format

A data frame with 17 observations on the following 2 variables. ======= A data frame with 17 observations on the following two variables. >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

hospital

a factor with levels Carolinas Med Ct, Duke Med Ct, Durham Regional, Forsyth Memorial, Frye Regional, High Point Regional, Memorial Mission, Mercy, Moore Regional, Moses Cone Memorial, NC Baptist, New Hanover Regional, Pitt Co. Memorial, Presbyterian, Rex, Univ of North Carolina, and Wake County
<<<<<<< HEAD
charge

a numeric vector
=======
charge

median charge for coronary bypass
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Bypass

Data for Exercises 5.104 and 6.43

<<<<<<< HEAD str(Bypass) attach(Bypass) EDA(charge) t.test(charge,conf.level=.90)$conf t.test(charge,mu=35000) detach(Bypass)

======= EDA(Bypass$charge) t.test(Bypass$charge, conf.level=.90)$conf t.test(Bypass$charge, mu = 35000)

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets