Insurance injury ratings of Chevrolet vehicles for 1990 and 1993 models

Data for Exercise 8.35

Chevy

Format

A data frame with 67 observations on two variables.

year

a factor with levels 1988-90 and 1991-93

frequency

a factor with levels much better than average, above average, average, below average, and much worse than average

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury

Examples

str(Chevy)

#> 'data.frame':	67 obs. of  2 variables:
#>  $ year     : Factor w/ 2 levels "1988-90","1991-93": 1 1 1 1 1 1 1 1 1 1 ...
#>  $ frequency: Factor w/ 5 levels "much better than average",..: 1 1 1 1 1 1 1 1 1 1 ...
#> 

T1 <- xtabs(~year + frequency, data = Chevy)
T1

#>          frequency
#> year      much better than average above average average below average
#>   1988-90                       16             5       5             3
#>   1991-93                       12             2      12             2
#>          frequency
#> year      much worse than average
#>   1988-90                       4
#>   1991-93                       6
#> 

chisq.test(T1)

Warning message:
Chi-squared approximation may be incorrect

#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  T1
#> X-squared = 5.3258, df = 4, p-value = 0.2555
#> 
#>