Percent of peak bone density of different aged children

Data for Exercise 9.7

Citrus

Format

A data frame with nine observations on the following two variables.

age

age of children

percent

percent peak bone density

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury

Examples

str(Citrus)

#> 'data.frame':	9 obs. of  2 variables:
#>  $ age    : int  2 4 6 8 10 12 14 16 18
#>  $ percent: int  43 49 51 56 63 71 82 91 95
#> 

model <- lm(percent ~ age, data = Citrus)
summary(model)

#> 
#> Call:
#> lm(formula = percent ~ age, data = Citrus)
#> 
#> Residuals:
#>     Min      1Q  Median      3Q     Max 
#> -3.9278 -2.6278  0.8222  2.7722  3.6722 
#> 
#> Coefficients:
#>             Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)  32.5278     2.4283   13.40 3.03e-06 ***
#> age           3.4250     0.2158   15.87 9.55e-07 ***
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> Residual standard error: 3.342 on 7 degrees of freedom
#> Multiple R-squared:  0.973,	Adjusted R-squared:  0.9691 
#> F-statistic:   252 on 1 and 7 DF,  p-value: 9.547e-07
#> 
#> 

anova(model)

#> Analysis of Variance Table
#> 
#> Response: percent
#>           Df  Sum Sq Mean Sq F value    Pr(>F)    
#> age        1 2815.35 2815.35     252 9.547e-07 ***
#> Residuals  7   78.21   11.17                      
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 

rm(model)