Commuting times for selected cities in 1980 and 1990

Format

A data frame with 39 observations on the following 3 variables.

City

a factor with levels Atlanta Baltimore Boston Buffalo Charlotte Chicago Cincinnati Cleveland Columbus Dallas Denver Detroit Hartford Houston Indianapolis Kansas City Los Angeles Miami Milwaukee Minneapolis New Orleans New York Norfolk Orlando Philadelphia Phoenix Pittsburgh Portland Providence Rochester Sacramento Salt Lake City San Antonio San Diego San Francisco Seattle St. Louis Tampa ======= A data frame with 39 observations on the following three variables.

city

a factor with levels Atlanta, Baltimore, Boston, Buffalo, Charlotte, Chicago, Cincinnati, Cleveland, Columbus, Dallas, Denver, Detroit, Hartford, Houston, Indianapolis, Kansas City, Los Angeles, Miami, Milwaukee, Minneapolis, New Orleans, New York, Norfolk, Orlando, Philadelphia, Phoenix, Pittsburgh, Portland, Providence, Rochester, Sacramento, Salt Lake City, San Antonio, San Diego, San Francisco, Seattle, St. Louis, Tampa, and >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051 Washington

year1980

urban commute times for the year 1980

year1990

urban commute times for the year 1990

Commute

Data for Exercises 1.13, and 7.85

<<<<<<< HEAD str(Commute) attach(Commute) stripchart(x=list(X1980,X1990),method="stack",pch=1,cex=2,col=c("red","blue"), group.names=c("1980","1990"),main="",xlab="minutes") title(main="Commute Time") boxplot(X1980,X1990,col=c("red","blue"),names=c("1980","1990"),horizontal=TRUE,las=1) library(lattice) commute <- stack(Commute) commute[1:5,] attach(commute) stripplot(ind~values,jitter=TRUE) dotplot(ind~values) bwplot(ind~values) remove(commute) detach(Commute)

======= commute <- stack(Commute) str(commute) stripplot(ind ~ values, data = commute, jitter = TRUE) dotplot(ind ~ values, data = commute) bwplot(ind ~ values, data = commute) stripchart(values ~ ind, data = commute, method = "stack", pch = 1, cex = 2, col = c("red", "blue"), group.names = c("1980", "1990"), main = "", xlab = "minutes") title(main = "Commute Time") boxplot(values ~ ind, data = commute, names=c("1980", "1990"), horizontal = TRUE, las = 1) rm(commute)

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets