Violent crime rates for the states in 1983 and 1993

Format

A data frame with 51 observations on the following 3 variables.

State

a factor with levels Alabama Alaska Arizona Arkansas California Colorado Connecticut DC Delaware Florida Georgia Hawaii Idaho Illinois Indiana Iowa Kansas Kentucky Louisiana Maine Maryland Massachusetts Michigan Minnesota Mississippi Missour Montana Nebraska Nevada New Hampshire New Jersey New Mexico New York North Carolina North Dakota Ohio Oklahoma Oregon Pennsylvania Rhode Island South Carolina South Dakota Tennessee Texas Utah Vermont Virginia Washington West Virginia Wisconsin Wyoming

X1983

a numeric vector

X1993

a numeric vector
======= A data frame with 102 observations on the following three variables.

state

a factor with levels Alabama, Alaska, Arizona, Arkansas, California, Colorado, Connecticut, DC, Delaware, Florida, Georgia, Hawaii, Idaho, Illinois, Indiana, Iowa, Kansas, Kentucky, Louisiana, Maine, Maryland, Massachusetts, Michigan, Minnesota, Mississippi, Missour, Montana, Nebraska, Nevada, New Hampshire, New Jersey, New Mexico, New York, North Carolina, North Dakota, Ohio, Oklahoma, Oregon, Pennsylvania, Rhode Island, South Carolina, South Dakota, Tennessee, Texas, Utah, Vermont, Virginia, Washington, West Virginia, Wisconsin, and Wyoming

year

a factor with levels 1983 and 1993

rate

crime rate per 100,000 inhabitants
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Crime

Data for Exercises 1.90, 2.32, 3.64, and 5.113

str(Crime) <<<<<<< HEAD attach(Crime) boxplot(X1983,X1993,names=c("1983","1993"),xlab="Year", ylab="Crime Rate per 100,000 Inhabitants",main="Problem 1.90") plot(X1983,X1993) detach(Crime)

======= boxplot(rate ~ year, data = Crime)

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets