Percent of bachelor's degrees awarded women in 1970 versus 1990

Format

A data frame with 11 observations on the following 3 variables.

Field

a factor with levels All fields Business Education Engineering Fine Arts Foreign Lng Health Life Sciences Physical Sci Psychology Social Science

X1970

a numeric vector

X1990

a numeric vector
======= A data frame with 1064 observations on the following two variables.

field

a factor with levels Health, Education, Foreign Language, Psychology, Fine Arts, Life Sciences, Business, Social Science, Physical Sciences, Engineering, and All Fields

awarded

a factor with levels 1970 and 1990
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Degree

Data for Exercise 2.75

str(Degree) <<<<<<< HEAD attach(Degree) Dmat <- as.matrix(Degree[,2:3]) rownames(Dmat) <- Field colnames(Dmat) <- c("1970","1990") Dmat barplot(t(Dmat),beside=TRUE,legend=TRUE,cex.names=.5) detach(Degree) remove(Dmat)

======= T1 <- xtabs(~field + awarded, data = Degree) T1 barplot(t(T1), beside = TRUE, col = c("red", "skyblue"), legend = colnames(T1))

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets