Parking tickets issued to diplomats

Format

A data frame with 10 observations on the following 4 variables.

Country

a factor with levels Brazil Bulgaria Egypt Indonesia Israel Nigeria Russia S. Korea Ukraine Venezuela

Number

a numeric vector

rate

a numeric vector

Code

a factor with levels Br Bu Eg In Is Ni Ru SK Uk Ve
======= A data frame with 10 observations on the following three variables.

country

a factor with levels Brazil Bulgaria Egypt Indonesia Israel Nigeria Russia S. Korea Ukraine Venezuela

number

total number of tickets

rate

number of tickets per vehicle per month
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Diplomat

Data for Exercises 1.14 and 1.37

str(Diplomat) <<<<<<< HEAD attach(Diplomat) par(mfrow=c(1,2)) names(Number) <- Country dotchart(Number,main="Number of Tickets",col="blue",pch=1) names(rate) <- Country dotchart(rate,main="Tickets/Vehicle/Month",col="red",pch=2) barplot(rate,col="pink") detach(Diplomat)

======= par(las = 2, mfrow = c(2, 2)) stripchart(number ~ country, data = Diplomat, pch = 19, col= "red", vertical = TRUE) stripchart(rate ~ country, data = Diplomat, pch = 19, col= "blue", vertical = TRUE) with(data = Diplomat, barplot(number, names.arg = country, col = "red")) with(data = Diplomat, barplot(rate, names.arg = country, col = "blue")) par(las = 1, mfrow = c(1, 1)) ## Not run: ------------------------------------ # ggplot2::ggplot(data = Diplomat, aes(x = reorder(country, number), # y = number)) + # geom_bar(stat = "identity", fill = "pink", color = "black") + # theme_bw() + labs(x = "", y = "Total Number of Tickets") # ggplot2::ggplot(data = Diplomat, aes(x = reorder(country, rate), # y = rate)) + # geom_bar(stat = "identity", fill = "pink", color = "black") + # theme_bw() + labs(x = "", y = "Tickets per vehicle per month") ## ---------------------------------------------

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets