Rankings of the favorite breeds of dogs

Format

A data frame with 20 observations on the following 5 variables.

Dog

a factor with levels Beagle Boxer Chihuahua Chow Dachshund Dalmatian Doberman Huskie Labrador Pomeranian Poodle Retriever Rotweiler Schnauzer Shepherd Shetland ShihTzu Spaniel Springer Yorkshire

X1992

a numeric vector

X1993

a numeric vector

X1997

a numeric vector

X1998

a numeric vector
======= A data frame with 20 observations on the following three variables.

breed

a factor with levels Beagle, Boxer, Chihuahua, Chow, Dachshund, Dalmatian, Doberman, Huskie, Labrador, Pomeranian, Poodle, Retriever, Rotweiler, Schnauzer, Shepherd, Shetland, ShihTzu, Spaniel, Springer, and Yorkshire

ranking

numeric ranking

year

a factor with levels 1992, 1993, 1997, and 1998
>>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Dogs

Data for Exercise 2.88

str(Dogs) <<<<<<< HEAD attach(Dogs) cor(Dogs[,2:5]) detach(Dogs)

======= cor(Dogs$ranking[Dogs$year == "1992"], Dogs$ranking[Dogs$year == "1993"]) cor(Dogs$ranking[Dogs$year == "1997"], Dogs$ranking[Dogs$year == "1998"]) ## Not run: ------------------------------------ # ggplot2::ggplot(data = Dogs, aes(x = reorder(breed, ranking), y = ranking)) + # geom_bar(stat = "identity") + # facet_grid(year ~. ) + # theme(axis.text.x = element_text(angle = 85, vjust = 0.5)) ## ---------------------------------------------

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

datasets