Advertised salaries offered general managers of major corporations in 1995

Data for Exercise 5.81

Manager

Format

A data frame/tibble with 26 observations on one variable

salary

random sample of advertised annual salaries of top executives (in dollars)

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.

Examples

stem(Manager$salary)

#> 
#>   The decimal point is 4 digit(s) to the right of the |
#> 
#>    4 | 5
#>    6 | 00605568
#>    8 | 55560558
#>   10 | 000
#>   12 | 00
#>   14 | 000
#>   16 | 
#>   18 | 5
#> 

SIGN.test(Manager$salary)

#> 
#> 	One-sample Sign-Test
#> 
#> data:  Manager$salary
#> s = 26, p-value = 2.98e-08
#> alternative hypothesis: true median is not equal to 0
#> 95 percent confidence interval:
#>  76629.78 99954.81
#> sample estimates:
#> median of x 
#>       87750 
#> 
#> Achieved and Interpolated Confidence Intervals: 
#> 
#>                   Conf.Level   L.E.pt    U.E.pt
#> Lower Achieved CI     0.9245 78000.00  99900.00
#> Interpolated CI       0.9500 76629.78  99954.81
#> Upper Achieved CI     0.9710 75500.00 100000.00
#>