Math proficiency and SAT scores by states

Data for Exercise 9.24, Example 9.1, and Example 9.6

Format

A data frame with 51 observations on the following 10 variables.

state1

a factor with levels Conn D.C. Del Ga Hawaii Ind Maine Mass Md N.C. N.H. N.J. N.Y. Ore Pa R.I. S.C. Va Vt

Sat.M1

a numeric vector

Profic1

a numeric vector

state2

a factor with levels Ala Alaska Ariz Ark Calif Colo Fla Idaho Ill Iowa Kan Ky La Mich Minn Miss Mo Mont N.D. N.M. Neb Nev Ohio Okla S.D. Tenn Texas Utah W.V. Wash Wis Wyo

Sat.M2

a numeric vector

Profic2

a numeric vector

state

a factor with levels Ala Alaska Ariz Ark Calif Colo Conn D.C. Del Fla Ga Hawaii Idaho Ill Ind Iowa Kan Ky La Maine Mass Md Mich Minn Miss Mo Mont N.C. N.D. N.H. N.J. N.M. N.Y. Neb Nev Ohio Okla Ore Pa R.I. S.C. S.D. Tenn Texas Utah Va Vt W.V. Wash Wis Wyo

Sat.M

a numeric vector

Profic

a numeric vector

Group

a numeric vector

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Examples

str(Mathpro)
attach(Mathpro)
model <- lm(Sat.M1~Profic1)
plot(Profic1,Sat.M1)
abline(model)
model
detach(Mathpro)
remove(model)

<<<<<<< HEAD
=======

Error: <text>:11:1: unexpected input
10: 
11: <<
    ^