Carbon monoxide emitted by smoke stacks of a manufacturer and a competitor

Data for Exercise 7.45

Format

A data frame with 10 observations on the following 2 variables.

manufac

a numeric vector

compet

a numeric vector

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury

Examples

str(Monoxide)

#> 'data.frame':	10 obs. of  2 variables:
#>  $ manufac: num  2.7 3.1 3.1 2.9 2.5 3.4 3.4 3.4 2.4 NA
#>  $ compet : num  3.7 3 3.5 3.8 2.8 3.5 3.4 3.6 2.7 3.7
#> 

attach(Monoxide)
t.test(manufac,compet)

#> 
#> 	Welch Two Sample t-test
#> 
#> data:  manufac and compet
#> t = -2.1187, df = 16.842, p-value = 0.04929
#> alternative hypothesis: true difference in means is not equal to 0
#> 95 percent confidence interval:
#>  -0.760895683 -0.001326539
#> sample estimates:
#> mean of x mean of y 
#>  2.988889  3.370000 
#> 
#> 

wilcox.test(manufac,compet)

Warning message:
cannot compute exact p-value with ties

#> 
#> 	Wilcoxon rank sum test with continuity correction
#> 
#> data:  manufac and compet
#> W = 19, p-value = 0.03617
#> alternative hypothesis: true location shift is not equal to 0
#> 
#> 

detach(Monoxide)