Salaries of members of the Baltimore Orioles baseball team

Data for Example 1.3

Format

A data frame with 27 observations on the following 3 variables.

first.name

a factor with levels Albert Arthur B.J. Brady Cal Charles dl-Delino dl-Scott Doug Harold Heathcliff Jeff Jesse Juan Lenny Mike Rich Ricky Scott Sidney Will Willis

last.name

a factor with levels Amaral Anderson Baines Belle Bones Bordick Clark Conine Deshields Erickson Fetters Garcia Guzman Johns Johnson Kamieniecki Mussina Orosco Otanez Ponson Reboulet Rhodes Ripken Jr. Slocumb Surhoff Timlin Webster

X1999salary

a numeric vector

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury

Examples

data(Orioles)