SAT scores, percent taking exam and state funding per student by state for 1994, 1995 and 1999

Data for Statistical Insight Chapter 9

Format

A data frame with 51 observations on the following 16 variables.

state

a factor with levels alabama alaska arizona arkansas california colorado connecticut delaware dist of columbia florida georgia hawaii idaho illinois indiana iowa kansas kentucky louisiana maine maryland massachusetts michigan minnesota mississippi missouri montana nebraska nevada new hampshire new jersey new mexico new york north carolina north dakota ohio oklahoma oregon pennsylvania rhode island south carolina south dakota tennessee texas utah vermont virginia washington west virginia wisconsin wyoming

verbal94

a numeric vector

math94

a numeric vector

total94

a numeric vector

percent94

a numeric vector

code94

a numeric vector

expend94

a numeric vector

verbal95

a numeric vector

math95

a numeric vector

total95

a numeric vector

verbal99

a numeric vector

math99

a numeric vector

total99

a numeric vector

percent99

a numeric vector

code99

a numeric vector

expend99

a numeric vector

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury

Examples

str(Sat)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	102 obs. of  7 variables:
#>  $ state  : chr  "alabama" "alaska" "arizona" "arkansas" ...
#>  $ verbal : int  482 434 443 417 413 456 426 428 406 413 ...
#>  $ math   : int  529 477 496 518 482 513 472 464 443 466 ...
#>  $ total  : int  1011 911 939 935 895 969 898 892 849 879 ...
#>  $ percent: int  8 49 26 6 46 28 80 68 53 49 ...
#>  $ expend : int  3616 8450 4381 4031 4746 5172 8017 6093 9549 5243 ...
#>  $ year   : Factor w/ 2 levels "1994","1999": 1 1 1 1 1 1 1 1 1 1 ...
#>  - attr(*, "spec")=List of 2
#>   ..$ cols   :List of 16
#>   .. ..$ state    : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_character" "collector"
#>   .. ..$ verbal94 : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ math94   : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ total94  : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ percent94: list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ code94   : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ expend94 : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ verbal95 : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ math95   : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ total95  : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ verbal99 : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ math99   : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ total99  : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ percent99: list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ code99   : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ expend99 : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   ..$ default: list()
#>   .. ..- attr(*, "class")= chr  "collector_guess" "collector"
#>   ..- attr(*, "class")= chr "col_spec"

attach(Sat)
pairs(Sat)

#> Error in pairs.default(Sat): non-numeric argument to 'pairs'

detach(Sat)