Convictions in spouse murder cases by gender

Data for Exercise 8.33

Spouse

Format

A data frame/tibble with 540 observations on two variables

result

a factor with levels not prosecuted, pleaded guilty, convicted, and acquited

spouse

a factor with levels husband and wife

Source

Bureau of Justice Statistics (September 1995), Spouse Murder Defendants in Large Urban Counties, Executive Summary, NCJ-156831.

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.

Examples

T1 <- xtabs(~result + spouse, data = Spouse)
T1

#>                 spouse
#> result           husband wife
#>   not prosecuted      35   35
#>   pleaded guilty     146   87
#>   convicted          130   69
#>   acquitted            7   31

chisq.test(T1)

#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  T1
#> X-squared = 32.765, df = 3, p-value = 3.61e-07
#> 

rm(T1)