Baseball team wins versus 7 independent variables for National league teams

Data for Exercise 9.23

Format

A data frame with 12 observations on the following 9 variables.

team

a factor with levels Atlanta Chicago Cincinnati Houston Los Angeles Montreal New York Philadelphia Pittsburgh San Diego San Francisco St. Louis

wins

a numeric vector

batavg

a numeric vector

rbi

a numeric vector

stole

a numeric vector

strkout

a numeric vector

caught

a numeric vector

errors

a numeric vector

era

a numeric vector

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury >>>>>>> 832da852208a5c5dc5d4420208e4ca0f661b9051

Examples

str(Wins)
attach(Wins)
plot(era,wins)
model <- lm(wins~era)
abline(model)
summary(model)
detach(Wins)
remove(model)

<<<<<<< HEAD
=======

Error: <text>:11:1: unexpected input
10: 
11: <<
    ^