Z-test

This function is based on the standard normal distribution and creates confidence intervals and tests hypotheses for both one and two sample problems.

z.test(x, y = NULL, alternative = "two.sided", mu = 0, sigma.x = NULL,
  sigma.y = NULL, conf.level = 0.95)

Arguments

x

numeric vector; NAs and Infs are allowed but will be removed.

y

numeric vector; NAs and Infs are allowed but will be removed.

alternative

character string, one of "greater", "less" or "two.sided", or the initial letter of each, indicating the specification of the alternative hypothesis. For one-sample tests, alternative refers to the true mean of the parent population in relation to the hypothesized value mu. For the standard two-sample tests, alternative refers to the difference between the true population mean for x and that for y, in relation to mu.

mu

a single number representing the value of the mean or difference in means specified by the null hypothesis

sigma.x

a single number representing the population standard deviation for x

sigma.y

a single number representing the population standard deviation for y

conf.level

confidence level for the returned confidence interval, restricted to lie between zero and one

Value

A list of class htest, containing the following components:

Details

If y is NULL, a one-sample z-test is carried out with x. If y is not NULL, a standard two-sample z-test is performed.

Null Hypothesis

For the one-sample z-test, the null hypothesis is that the mean of the population from which x is drawn is mu. For the standard two-sample z-tests, the null hypothesis is that the population mean for x less that for y is mu.

The alternative hypothesis in each case indicates the direction of divergence of the population mean for x (or difference of means for x and y) from mu (i.e., "greater", "less", "two.sided").

References

Kitchens, L.J. (2003). Basic Statistics and Data Analysis. Duxbury.

Hogg, R. V. and Craig, A. T. (1970). Introduction to Mathematical Statistics, 3rd ed. Toronto, Canada: Macmillan.

Mood, A. M., Graybill, F. A. and Boes, D. C. (1974). Introduction to the Theory of Statistics, 3rd ed. New York: McGraw-Hill.

Snedecor, G. W. and Cochran, W. G. (1980). Statistical Methods, 7th ed. Ames, Iowa: Iowa State University Press.

See also

zsum.test, tsum.test

Examples

x <- rnorm(12)
z.test(x,sigma.x=1)

#> 
#> 	One-sample z-Test
#> 
#> data:  x
#> z = 0.64861, p-value = 0.5166
#> alternative hypothesis: true mean is not equal to 0
#> 95 percent confidence interval:
#>  -0.3785552  0.7530305
#> sample estimates:
#> mean of x 
#> 0.1872376 
#> 
#> 

        # Two-sided one-sample z-test where the assumed value for
        # sigma.x is one. The null hypothesis is that the population
        # mean for 'x' is zero. The alternative hypothesis states
        # that it is either greater or less than zero. A confidence
        # interval for the population mean will be computed.

x <- c(7.8, 6.6, 6.5, 7.4, 7.3, 7., 6.4, 7.1, 6.7, 7.6, 6.8)
y <- c(4.5, 5.4, 6.1, 6.1, 5.4, 5., 4.1, 5.5)
z.test(x, sigma.x=0.5, y, sigma.y=0.5, mu=2)

#> 
#> 	Two-sample z-Test
#> 
#> data:  x and y
#> z = -1.0516, p-value = 0.293
#> alternative hypothesis: true difference in means is not equal to 2
#> 95 percent confidence interval:
#>  1.300323 2.211040
#> sample estimates:
#> mean of x mean of y 
#>  7.018182  5.262500 
#> 
#> 

        # Two-sided standard two-sample z-test where both sigma.x
        # and sigma.y are both assumed to equal 0.5. The null hypothesis
        # is that the population mean for 'x' less that for 'y' is 2.
        # The alternative hypothesis is that this difference is not 2.
        # A confidence interval for the true difference will be computed.

z.test(x, sigma.x=0.5, y, sigma.y=0.5, conf.level=0.90)

#> 
#> 	Two-sample z-Test
#> 
#> data:  x and y
#> z = 7.5568, p-value = 4.13e-14
#> alternative hypothesis: true difference in means is not equal to 0
#> 90 percent confidence interval:
#>  1.373533 2.137831
#> sample estimates:
#> mean of x mean of y 
#>  7.018182  5.262500 
#> 
#> 

        # Two-sided standard two-sample z-test where both sigma.x and
        # sigma.y are both assumed to equal 0.5. The null hypothesis
        # is that the population mean for 'x' less that for 'y' is zero.
        # The alternative hypothesis is that this difference is not
        # zero.  A 90% confidence interval for the true difference will
        # be computed.
rm(x, y)

Author
Alan T. Arnholt