Abortion rate by region of country

Data used in Exercise 8.51

Abortion

Format

A data frame/tibble with 51 observations on the following 10 variables:

state

a character variable with values alabama, alaska, arizona, arkansas, california, colorado, connecticut, delaware, dist of columbia, florida, georgia, hawaii, idaho, illinois, indiana, iowa, kansas, kentucky, louisiana, maine, maryland, massachusetts, michigan, minnesota, mississippi, missouri, montana, nebraska, nevada, new hampshire, new jersey, new mexico, new york, north carolina, north dakota, ohio, oklahoma, oregon, pennsylvania, rhode island, south carolina, south dakota, tennessee, texas, utah, vermont, virginia, washington, west virginia, wisconsin, and wyoming

region

a character variable with values midwest northeast south west

regcode

a numeric vector

rate1988

a numeric vector

rate1992

a numeric vector

rate1996

a numeric vector

provide1988

a numeric vector

provide1992

a numeric vector

lowhigh

a numeric vector

rate

a factor with levels Low and High

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.

Examples

T1 <- xtabs(~region + rate, data = Abortion)
T1
#>            rate
#> region      high low
#>   midwest      3   9
#>   northeast    8   3
#>   south        6   9
#>   west         7   6
chisq.test(T1)
#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  T1
#> X-squared = 5.7932, df = 3, p-value = 0.1221
#>