Rankings of the favorite breeds of dogs

Data for Exercise 2.88

Dogs

Format

A data frame with 20 observations on the following three variables.

breed

a factor with levels Beagle, Boxer, Chihuahua, Chow, Dachshund, Dalmatian, Doberman, Huskie, Labrador, Pomeranian, Poodle, Retriever, Rotweiler, Schnauzer, Shepherd, Shetland, ShihTzu, Spaniel, Springer, and Yorkshire

ranking

numeric ranking

year

a factor with levels 1992, 1993, 1997, and 1998

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury

Examples

str(Dogs)

#> 'data.frame':	80 obs. of  3 variables:
#>  $ breed  : Factor w/ 20 levels "Beagle","Boxer",..: 9 13 15 18 12 11 1 5 6 16 ...
#>  $ ranking: num  1 2 4 3 6 5 7 8 15 9 ...
#>  $ year   : Factor w/ 4 levels "1992","1993",..: 1 1 1 1 1 1 1 1 1 1 ...
#> 

cor(Dogs$ranking[Dogs$year == "1992"], Dogs$ranking[Dogs$year == "1993"])

#> [1] 0.9413534
#> 

cor(Dogs$ranking[Dogs$year == "1997"], Dogs$ranking[Dogs$year == "1998"])

#> [1] 0.9005587
#> 

## Not run: ------------------------------------
# ggplot2::ggplot(data = Dogs, aes(x = reorder(breed, ranking), y = ranking)) + 
#      geom_bar(stat = "identity") + 
#      facet_grid(year ~. ) + 
#      theme(axis.text.x  = element_text(angle = 85, vjust = 0.5)) 
## ---------------------------------------------