Length and number of fish caught with small and large mesh codend

Data for Exercises 5.83, 5.119, and 7.29

Fish

Format

A data frame/tibble with 1534 observations on two variables

codend

a character variable with values smallmesh and largemesh

length

length of the fish measured in centimeters

Source

R. Millar, “Estimating the Size - Selectivity of Fishing Gear by Conditioning on the Total Catch,” Journal of the American Statistical Association, 87 (1992), 962 - 968.

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Pacific Grove, CA: Brooks/Cole, a division of Thomson Learning.

Examples

tapply(Fish$length, Fish$codend, median, na.rm = TRUE)
#> largemesh smallmesh 
#>        34        33 
SIGN.test(Fish$length[Fish$codend == "smallmesh"], conf.level = 0.99)
#> 
#> 	One-sample Sign-Test
#> 
#> data:  Fish$length[Fish$codend == "smallmesh"]
#> s = 739, p-value < 2.2e-16
#> alternative hypothesis: true median is not equal to 0
#> 99 percent confidence interval:
#>  33 34
#> sample estimates:
#> median of x 
#>          33 
#> 
#> Achieved and Interpolated Confidence Intervals: 
#> 
#>                   Conf.Level L.E.pt U.E.pt
#> Lower Achieved CI     0.9877     33     34
#> Interpolated CI       0.9900     33     34
#> Upper Achieved CI     0.9900     33     34
#> 
if (FALSE) {
dplyr::group_by(Fish, codend) %>%
         summarize(MEDIAN = median(length, na.rm = TRUE))
}