Percent of vehicles passing inspection by type inspection station

Data for Exercise 5.89

Format

A data frame with 6 observations on the following 4 variables.

Type

a factor with levels auto inspection auto repair car care center gas station new car dealer tire store

less70

a numeric vector

X70.85

a numeric vector

great85

a numeric vector

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury

Examples

str(Inspect)

#> 'data.frame':	6 obs. of  4 variables:
#>  $ Type   : Factor w/ 6 levels "auto inspection",..: 1 2 6 4 3 5
#>  $ less70 : int  1 16 16 19 4 1
#>  $ X70.85 : int  8 13 11 21 4 5
#>  $ great85: int  4 7 6 6 4 28

attach(Inspect)
Inspect

#>              Type less70 X70.85 great85
#> 1 auto inspection      1      8       4
#> 2     auto repair     16     13       7
#> 3      tire store     16     11       6
#> 4     gas station     19     21       6
#> 5 car care center      4      4       4
#> 6  new car dealer      1      5      28

chisq.test(Inspect[,2:4])

#> Warning: Chi-squared approximation may be incorrect

#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  Inspect[, 2:4]
#> X-squared = 61.027, df = 10, p-value = 2.316e-09
#> 

detach(Inspect)