Math proficiency and SAT scores by states

Data for Exercise 9.24, Example 9.1, and Example 9.6

Format

A data frame with 51 observations on the following 10 variables.

state1

a factor with levels Conn D.C. Del Ga Hawaii Ind Maine Mass Md N.C. N.H. N.J. N.Y. Ore Pa R.I. S.C. Va Vt

Sat.M1

a numeric vector

Profic1

a numeric vector

state2

a factor with levels Ala Alaska Ariz Ark Calif Colo Fla Idaho Ill Iowa Kan Ky La Mich Minn Miss Mo Mont N.D. N.M. Neb Nev Ohio Okla S.D. Tenn Texas Utah W.V. Wash Wis Wyo

Sat.M2

a numeric vector

Profic2

a numeric vector

state

a factor with levels Ala Alaska Ariz Ark Calif Colo Conn D.C. Del Fla Ga Hawaii Idaho Ill Ind Iowa Kan Ky La Maine Mass Md Mich Minn Miss Mo Mont N.C. N.D. N.H. N.J. N.M. N.Y. Neb Nev Ohio Okla Ore Pa R.I. S.C. S.D. Tenn Texas Utah Va Vt W.V. Wash Wis Wyo

Sat.M

a numeric vector

Profic

a numeric vector

Group

a numeric vector

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury

Examples

str(Mathpro)

#> Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	51 obs. of  4 variables:
#>  $ state   : chr  "Conn" "Del" "D.C." "Ga" ...
#>  $ sat_math: int  472 464 443 446 480 466 463 479 475 486 ...
#>  $ profic  : num  23.1 16.1 2.8 14.7 12.5 17.5 NA 17.1 NA 22.5 ...
#>  $ group   : int  1 1 1 1 1 1 1 1 1 1 ...
#>  - attr(*, "spec")=List of 2
#>   ..$ cols   :List of 10
#>   .. ..$ state1 : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_character" "collector"
#>   .. ..$ Sat-M1 : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ Profic1: list()
#>   .. .. ..- attr(*, "class")= chr  "collector_double" "collector"
#>   .. ..$ state2 : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_character" "collector"
#>   .. ..$ Sat-M2 : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ Profic2: list()
#>   .. .. ..- attr(*, "class")= chr  "collector_double" "collector"
#>   .. ..$ state  : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_character" "collector"
#>   .. ..$ Sat-M  : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   .. ..$ Profic : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_double" "collector"
#>   .. ..$ Group  : list()
#>   .. .. ..- attr(*, "class")= chr  "collector_integer" "collector"
#>   ..$ default: list()
#>   .. ..- attr(*, "class")= chr  "collector_guess" "collector"
#>   ..- attr(*, "class")= chr "col_spec"

attach(Mathpro)
model <- lm(Sat.M1~Profic1)

#> Error in eval(predvars, data, env): object 'Sat.M1' not found

plot(Profic1,Sat.M1)

#> Error in plot(Profic1, Sat.M1): object 'Profic1' not found

abline(model)

#> Error in int_abline(a = a, b = b, h = h, v = v, untf = untf, ...): plot.new has not been called yet

model

#> 
#> Call:
#> lm(formula = number ~ gallons, data = Wastewat)
#> 
#> Coefficients:
#> (Intercept)      gallons  
#>       3.590        3.791  
#> 

detach(Mathpro)
remove(model)

#> Warning: object 'model' not found