Sign Test

This function will test a hypothesis based on the sign test and reports linearly interpolated confidence intervals for one sample problems.

SIGN.test(x, y = NULL, md = 0, alternative = "two.sided",
  conf.level = 0.95)

Arguments

x

numeric vector; NAs and Infs are allowed but will be removed.

y

optional numeric vector; NAs and Infs are allowed but will be removed.

md

a single number representing the value of the population median specified by the null hypothesis

alternative

is a character string, one of "greater", "less", or "two.sided", or the initial letter of each, indicating the specification of the alternative hypothesis. For one-sample tests, alternative refers to the true median of the parent population in relation to the hypothesized value of the median.

conf.level

confidence level for the returned confidence interval, restricted to lie between zero and one

Value

A list of class htest, containing the following components:

Details

Computes a “Dependent-samples Sign-Test” if both x and y are provided. If only x is provided, computes the “Sign-Test”.

Note

The reported confidence interval is based on linear interpolation. The lower and upper confidence levels are exact.

Null Hypothesis

For the one-sample sign-test, the null hypothesis is that the median of the population from which x is drawn is md. For the two-sample dependent case, the null hypothesis is that the median for the differences of the populations from which x and y are drawn is md. The alternative hypothesis indicates the direction of divergence of the population median for x from md (i.e., "greater", "less", "two.sided".)

References

Gibbons, J.D. and Chakraborti, S. (1992). Nonparametric Statistical Inference. Marcel Dekker Inc., New York.

Kitchens, L.J.(2003). Basic Statistics and Data Analysis. Duxbury.

Conover, W. J. (1980). Practical Nonparametric Statistics, 2nd ed. Wiley, New York.

Lehmann, E. L. (1975). Nonparametrics: Statistical Methods Based on Ranks. Holden and Day, San Francisco.

See also

z.test, zsum.test, tsum.test

Examples

x <- c(7.8, 6.6, 6.5, 7.4, 7.3, 7., 6.4, 7.1, 6.7, 7.6, 6.8)
SIGN.test(x,md=6.5)

#> 
#> 	One-sample Sign-Test
#> 
#> data:  x
#> s = 9, p-value = 0.02148
#> alternative hypothesis: true median is not equal to 6.5
#> 95 percent confidence interval:
#>  6.571273 7.457455
#> sample estimates:
#> median of x 
#>           7 
#> 

#>                   Conf.Level L.E.pt U.E.pt
#> Lower Achieved CI     0.9346 6.6000 7.4000
#> Interpolated CI       0.9500 6.5713 7.4575
#> Upper Achieved CI     0.9883 6.5000 7.6000

        # Computes two-sided sign-test for the null hypothesis 
        # that the population median for 'x' is 6.5. The alternative 
        # hypothesis is that the median is not 6.5. An interpolated 95% 
        # confidence interval for the population median will be computed.

reaction <- c(14.3, 13.7, 15.4, 14.7, 12.4, 13.1, 9.2, 14.2,
              14.4, 15.8, 11.3, 15.0)
SIGN.test(reaction, md=15, alternative="less")

#> 
#> 	One-sample Sign-Test
#> 
#> data:  reaction
#> s = 2, p-value = 0.03271
#> alternative hypothesis: true median is less than 15
#> 95 percent confidence interval:
#>      -Inf 14.82845
#> sample estimates:
#> median of x 
#>       14.25 
#> 

#>                   Conf.Level L.E.pt  U.E.pt
#> Lower Achieved CI     0.9270   -Inf 14.7000
#> Interpolated CI       0.9500   -Inf 14.8285
#> Upper Achieved CI     0.9807   -Inf 15.0000

        # Data from Example 6.11 page 330 of Kitchens BSDA.  
        # Computes one-sided sign-test for the null hypothesis 
        # that the population median is 15.  The alternative 
        # hypothesis is that the median is less than 15.  
        # An interpolated upper 95% upper bound for the population 
        # median will be computed.