Margin of victory in Superbowls I to XXXV

Data for Exercise 1.54

Format

A data frame with 35 observations on the following 5 variables.

Winning.team

a factor with levels Baltimore Colts Baltimore Ravens Chicago Bears Dallas Cowboys Denver Broncos Green Bay Packers Kansas City Chiefs Los Angeles Raiders Miami Dolphins New York Giants New York Jets Oakland Raiders Pittsburgh Steelers San Francisco 49ers St Louis Rams Washington Redskins

winner.score

a numeric vector

Losing.team

a factor with levels Atlanta Falcons Baltimore Colts Buffalo Bills Cincinnati Bengals Dallas Cowboys Denver Broncos Green Bay Packers Kansas City Chiefs Los Angeles Rams Miami Dolphins Minnesota Vikings New England Patriots New York Giants Oakland Raiders Philadelphia Eagles Pittsburgh Steelers San Diego Chargers Tennessee Titans Washington Redskins

loser.score

a numeric vector

margin

a numeric vector

References

Kitchens, L. J. (2003) Basic Statistics and Data Analysis. Duxbury

Examples

str(Superbowl)

#> 'data.frame':	35 obs. of  5 variables:
#>  $ Winning.team: Factor w/ 16 levels "Baltimore Colts",..: 6 6 11 7 1 4 9 9 13 13 ...
#>  $ winner.score: int  35 33 16 23 16 24 14 24 16 21 ...
#>  $ Losing.team : Factor w/ 19 levels "Atlanta Falcons",..: 8 14 2 11 5 10 19 11 11 5 ...
#>  $ loser.score : int  10 14 7 7 13 3 7 7 6 17 ...
#>  $ margin      : int  25 19 9 16 3 21 7 17 10 4 ...

attach(Superbowl)
stem(margin)

#> 
#>   The decimal point is 1 digit(s) to the right of the |
#> 
#>   0 | 13444
#>   0 | 57779
#>   1 | 000234
#>   1 | 567777899
#>   2 | 123
#>   2 | 579
#>   3 | 2
#>   3 | 56
#>   4 | 
#>   4 | 5
#> 

detach(Superbowl)