Required Sample Size

Function to determine required sample size to be within a given margin of error.

nsize(b, sigma = NULL, p = 0.5, conf.level = 0.95, type = "mu")

Arguments

b

the desired bound.

sigma

population standard deviation. Not required if using type "pi".

p

estimate for the population proportion of successes. Not required if using type "mu".

conf.level

confidence level for the problem, restricted to lie between zero and one.

type

character string, one of "mu" or "pi", or just the initial letter of each, indicating the appropriate parameter. Default value is "mu".

Value

Returns required sample size.

Details

Answer is based on a normal approximation when using type "pi".

Examples

nsize(b=.03, p=708/1200, conf.level=.90, type="pi")

#> 
#> The required sample size (n) to estimate the population 
#> proportion of successes with a 0.9 confidence interval 
#> so that the margin of error is no more than 0.03 is 728 . 
#> 
#> 
#> 

    # Returns the required sample size (n) to estimate the population 
    # proportion of successes with a 0.9 confidence interval 
    # so that the margin of error is no more than 0.03 when the
    # estimate of the population propotion of successes is 708/1200.
    # This is problem 5.38 on page 257 of Kitchen's BSDA.

nsize(b=.15, sigma=.31, conf.level=.90, type="mu")

#> 
#> The required sample size (n) to estimate the population 
#> mean with a 0.9 confidence interval so that the margin 
#> of error is no more than 0.15 is 12 . 
#> 
#> 
#> 

    # Returns the required sample size (n) to estimate the population 
    # mean with a 0.9 confidence interval so that the margin 
    # of error is no more than 0.15.  This is Example 5.17 on page
    # 261 of Kitchen's BSDA.

Author
Alan T. Arnholt